Added DatesListData; js_only.
