+++ assets/prototype/domain/eventEditor/datetimes/dateList/AddNewDateButton.js
@@ -2,6 +2,14 @@ import { useState } from '@wordpress/element';
 import AddNewDateModal from './AddNewDateModal';
 import { EspressoButton } from '../../../../../ZZZ/components/ui';
 
+const btnRowStyle = {
+	display: 'flex',
+	flexFlow: 'row wrap',
+	justifyContent: 'flex-end',
+	margin: '0 0 2rem',
+	width: '100%',
+};
+
 const AddNewDateButton = () => {
 	const [isOpen, setIsOpen] = useState(false);
 	const handleOpen = (e) => {
@@ -16,10 +24,10 @@ const AddNewDateButton = () => {
 	};
 
 	return (
-		<>
+		<div style={btnRowStyle}>
 			<EspressoButton icon={'calendar'} buttonText={'Add New Date'} onClick={handleOpen} />
 			<AddNewDateModal handleClose={handleClose} isOpen={isOpen} />
-		</>
+		</div>
 	);
 };
